+++ internal/provider/monitor_report_data_source.go
@@ -163,7 +163,7 @@ func (d *MonitorReportDataSource) Read(ctx context.Context, req datasource.ReadR
 
 	report, err := d.client.GetMonitorReport(ctx, config.ID.ValueString(), from, to)
 	if err != nil {
-		resp.Diagnostics.Append(newReadError("Monitor Report", config.ID.ValueString(), err))
+		resp.Diagnostics.Append(NewReadErrorWithContext("Monitor Report", config.ID.ValueString(), err))
 		return
 	}